LongResponseException added and some fixes with ZH
