+++ src/main/java/telegrambot/controller/WebhookBotController.java
@@ -4,18 +4,20 @@ package telegrambot.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import telegrambot.config.telegram.TelegramWalletBot;
 
+@RequestMapping(value = "/webhook")
 @RestController
 @RequiredArgsConstructor
 public class WebhookBotController {
 
     private final TelegramWalletBot telegramWalletBot;
 
-    @PostMapping
+    @PostMapping(value = "/general")
     public SendMessage receiveUpdate(@RequestBody Update update) {
         return telegramWalletBot.onWebhookUpdateReceived(update);
     }
