+++ src/main/java/telegrambot/execurors/card/ConfirmCreateCardExecutor.java
@@ -6,11 +6,11 @@ import org.springframework.transaction.annotation.Transactional;
 import telegrambot.config.interceptor.UserDataContextHolder;
 import telegrambot.execurors.AbstractCommandExecutor;
 import telegrambot.model.Card;
+import telegrambot.model.drafts.CardDraft;
 import telegrambot.model.enums.DraftStatus;
 import telegrambot.model.util.Command;
 import telegrambot.model.util.CurrentCondition;
 import telegrambot.model.util.State;
-import telegrambot.model.util.drafts.CardDraft;
 import telegrambot.service.card.CardService;
 import telegrambot.service.card_draft.CardDraftService;
 import telegrambot.service.command.CommandService;
