+++ src/main/java/telegrambot/execurors/card/CreateCardExecutor.java
@@ -4,9 +4,9 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import telegrambot.config.interceptor.UserDataContextHolder;
 import telegrambot.execurors.AbstractCommandExecutor;
+import telegrambot.model.drafts.CardDraft;
 import telegrambot.model.util.CurrentCondition;
 import telegrambot.model.util.MsgFromStateHistory;
-import telegrambot.model.util.drafts.CardDraft;
 import telegrambot.service.card.CardService;
 import telegrambot.service.card_draft.CardDraftService;
 import telegrambot.service.command.CommandService;
