+++ src/main/java/telegrambot/execurors/transaction/ConfirmCreateTransactionExecutor.java
@@ -7,12 +7,12 @@ import telegrambot.config.interceptor.UserDataContextHolder;
 import telegrambot.execurors.AbstractCommandExecutor;
 import telegrambot.model.Card;
 import telegrambot.model.Transaction;
+import telegrambot.model.drafts.TransactionDraft;
 import telegrambot.model.enums.DraftStatus;
 import telegrambot.model.enums.TransactionTypeEnum;
 import telegrambot.model.util.Command;
 import telegrambot.model.util.CurrentCondition;
 import telegrambot.model.util.State;
-import telegrambot.model.util.drafts.TransactionDraft;
 import telegrambot.service.card.CardService;
 import telegrambot.service.command.CommandService;
 import telegrambot.service.current_condition.CurrentConditionService;
