+++ src/main/java/telegrambot/execurors/transaction/CreateTransactionExecutor.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Component;
 import telegrambot.config.interceptor.UserDataContextHolder;
 import telegrambot.execurors.AbstractCommandExecutor;
 import telegrambot.model.Card;
+import telegrambot.model.drafts.TransactionDraft;
 import telegrambot.model.enums.DraftStatus;
 import telegrambot.model.enums.StateEnum;
 import telegrambot.model.enums.TransactionTypeEnum;
@@ -12,7 +13,6 @@ import telegrambot.model.util.Command;
 import telegrambot.model.util.CurrentCondition;
 import telegrambot.model.util.MsgFromStateHistory;
 import telegrambot.model.util.State;
-import telegrambot.model.util.drafts.TransactionDraft;
 import telegrambot.service.card.CardService;
 import telegrambot.service.command.CommandService;
 import telegrambot.service.current_condition.CurrentConditionService;
