+++ src/main/java/telegrambot/repository/CardRepository.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 @Repository
 public interface CardRepository extends JpaRepository<Card, Long> {
 
-    @Query(value = "SELECT c.id, c.name, c.balance  FROM cards c WHERE c.name=:name", nativeQuery = true)
+    @Query(value = "SELECT c.id, c.name, c.balance  FROM cards c WHERE c.name = :name", nativeQuery = true)
     Optional<Card> getByName(@Param("name") String name);
 
     @Transactional
