+++ src/main/java/telegrambot/repository/draft/CardDraftRepository.java
@@ -1,4 +1,4 @@
-package telegrambot.repository.util;
+package telegrambot.repository.draft;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import telegrambot.model.util.drafts.CardDraft;
+import telegrambot.model.drafts.CardDraft;
 
 import java.math.BigDecimal;
 import java.util.Optional;
@@ -17,11 +17,6 @@ public interface CardDraftRepository extends JpaRepository<CardDraft, Long> {
     @Modifying
     @Query(value = "UPDATE card_draft  SET name = :name", nativeQuery = true)
     int updateName(@Param("name") String name);
-
-    @Transactional
-    @Modifying
-    @Query(value = "UPDATE card_draft SET balance = :balance", nativeQuery = true)
-    int updateBalance(@Param("balance") BigDecimal balance);
 
     @Transactional
     @Modifying
@@ -38,9 +33,6 @@ public interface CardDraftRepository extends JpaRepository<CardDraft, Long> {
     @Modifying
     @Query(value = "INSERT INTO card_draft(id, status) VALUES (2, 'BUILDING')", nativeQuery = true)
     int createFirstDraft();
-
-    @Query(value = "SELECT c.id, c.name, c.balance, c.status FROM card_draft AS c WHERE c.id=:id", nativeQuery = true)
-    CardDraft getById(@Param("id") Long id);
 
     @Query(value = "SELECT c.id, c.name, c.balance, c.status FROM card_draft AS c LIMIT 1", nativeQuery = true)
     Optional<CardDraft> getFirstDraft();
