+++ src/main/java/telegrambot/service/card_draft/CardDraftService.java
@@ -1,7 +1,7 @@
 package telegrambot.service.card_draft;
 
+import telegrambot.model.drafts.CardDraft;
 import telegrambot.model.enums.DraftStatus;
-import telegrambot.model.util.drafts.CardDraft;
 
 import java.math.BigDecimal;
