+++ src/main/java/telegrambot/service/transaction_draft/TransactionDraftServiceImpl.java
@@ -2,8 +2,8 @@ package telegrambot.service.transaction_draft;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
-import telegrambot.model.util.drafts.TransactionDraft;
-import telegrambot.repository.util.TransactionDraftRepository;
+import telegrambot.model.drafts.TransactionDraft;
+import telegrambot.repository.draft.TransactionDraftRepository;
 
 import java.math.BigDecimal;
